Remove `UpdatePartnersCommand` and `UpdateWorkflow`, integrate partner update logic into `ScrapeWorkflow` and `ScrapePartnersCommand` for improved modularity and maintainability. Update ScrapeConfig to support specific partner updates. Streamline progress tracking and error handling.
