+++ src/Bitrix24Partners/Console/ScrapePartnersCommand.php
@@ -19,7 +19,7 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 
 #[AsCommand(
     name: 'partners:scrape',
-    description: 'Парсит партнеров Bitrix24 и сохраняет данные в CSV'
+    description: 'Парсит партнеров Bitrix24 и сохраняет данные в CSV. С --partner-ids — обновляет конкретных партнёров.'
 )]
 class ScrapePartnersCommand extends Command
 {
@@ -35,7 +35,7 @@ class ScrapePartnersCommand extends Command
 
     public function __construct(
         private readonly LoggerInterface $logger,
-        private readonly ScrapeWorkflow $workflow,
+        private readonly ScrapeWorkflow $scrapeWorkflow,
     ) {
         parent::__construct();
     }
@@ -49,6 +49,7 @@ class ScrapePartnersCommand extends Command
             ->addOption('catalog-page-delay', null, InputOption::VALUE_REQUIRED, 'Задержка между страницами каталога (сек)', (string) self::DEFAULT_CATALOG_PAGE_DELAY)
             ->addOption('partner-detail-delay', null, InputOption::VALUE_REQUIRED, 'Задержка между карточками партнёров (сек)', (string) self::DEFAULT_PARTNER_DETAIL_DELAY)
             ->addOption('insecure', null, InputOption::VALUE_NONE, 'Отключить проверку SSL (для dev)')
+            ->addOption('partner-ids', null, InputOption::VALUE_REQUIRED, 'ID партнёров через запятую (режим обновления)', '')
             ->addOption('resume', null, InputOption::VALUE_NONE, 'Продолжить с места обрыва (из state-файла)')
             ->addOption('full-refresh', null, InputOption::VALUE_NONE, 'Перечитать всех с сайта → перезаписать CSV')
         ;
@@ -69,14 +70,23 @@ class ScrapePartnersCommand extends Command
             $this->io->text(sprintf('Zone: %s', $config->zone->value));
             $this->io->text(sprintf('Base URL: %s', $config->baseUrl));
             $this->io->text(sprintf('Output file: %s', $config->outputFile));
-            $this->io->text(sprintf('Catalog page delay: %d sec', $config->catalogPageDelay));
             $this->io->text(sprintf('Partner detail delay: %d sec', $config->partnerDetailDelay));
             $this->io->text(sprintf('Insecure: %s', $config->insecure ? 'yes' : 'no'));
-            $this->io->text(sprintf('Resume: %s', $config->resume ? 'yes' : 'no'));
-            $this->io->text(sprintf('Full refresh: %s', $config->fullRefresh ? 'yes' : 'no'));
+
+            if ($config->isUpdateMode()) {
+                $this->io->text(sprintf('Partner IDs: %s', implode(', ', $config->partnerIds)));
+            } else {
+                $this->io->text(sprintf('Catalog page delay: %d sec', $config->catalogPageDelay));
+                $this->io->text(sprintf('Resume: %s', $config->resume ? 'yes' : 'no'));
+                $this->io->text(sprintf('Full refresh: %s', $config->fullRefresh ? 'yes' : 'no'));
+            }
         }
 
         try {
+            if ($config->isUpdateMode()) {
+                return $this->executeUpdate($config);
+            }
+
             return $this->executeFullScrape($config);
         } catch (\Throwable $throwable) {
             $this->logger->error('Ошибка: '.$throwable->getMessage());
@@ -100,16 +110,32 @@ class ScrapePartnersCommand extends Command
             return null;
         }
 
-        $catalogPageDelay = (int) $input->getOption('catalog-page-delay');
-        if ($catalogPageDelay <= 0) {
-            $this->io->error('catalog-page-delay must be greater than 0');
+        $partnerDetailDelay = (int) $input->getOption('partner-detail-delay');
+        if ($partnerDetailDelay <= 0) {
+            $this->io->error('partner-detail-delay must be greater than 0');
 
             return null;
         }
 
-        $partnerDetailDelay = (int) $input->getOption('partner-detail-delay');
-        if ($partnerDetailDelay <= 0) {
-            $this->io->error('partner-detail-delay must be greater than 0');
+        $partnerIds = null;
+        $partnerIdsRaw = $input->getOption('partner-ids');
+        if ('' !== $partnerIdsRaw) {
+            $partnerIds = [];
+            $parts = array_map('trim', explode(',', (string) $partnerIdsRaw));
+            foreach ($parts as $part) {
+                if (!ctype_digit($part)) {
+                    $this->io->error(sprintf('Невалидный ID партнёра: "%s". Ожидается положительное число.', $part));
+
+                    return null;
+                }
+
+                $partnerIds[] = (int) $part;
+            }
+        }
+
+        $catalogPageDelay = (int) $input->getOption('catalog-page-delay');
+        if (null === $partnerIds && $catalogPageDelay <= 0) {
+            $this->io->error('catalog-page-delay must be greater than 0');
 
             return null;
         }
@@ -122,7 +148,29 @@ class ScrapePartnersCommand extends Command
             insecure: (bool) $input->getOption('insecure'),
             resume: (bool) $input->getOption('resume'),
             fullRefresh: (bool) $input->getOption('full-refresh'),
+            partnerIds: $partnerIds,
         );
+    }
+
+    private function executeUpdate(ScrapeConfig $config): int
+    {
+        if ($this->output->getVerbosity() >= OutputInterface::VERBOSITY_NORMAL) {
+            $this->io->text(sprintf('Обновление %d партнёров...', count($config->partnerIds)));
+        }
+
+        $progressBar = $this->createUpdateProgressBar(count($config->partnerIds));
+
+        $onProgress = function (string $event, int $value) use ($progressBar): void {
+            match ($event) {
+                'partner_start' => $progressBar?->setMessage((string) $value, 'partner'),
+                'partner_advance' => $progressBar?->advance(),
+                default => null,
+            };
+        };
+
+        $result = $this->scrapeWorkflow->runUpdate($config, $onProgress);
+
+        return $this->finishUpdate($progressBar, $result);
     }
 
     private function executeFullScrape(ScrapeConfig $config): int
@@ -137,7 +185,7 @@ class ScrapePartnersCommand extends Command
             ? fn (string $message) => $this->io->text($message)
             : null;
 
-        $context = $this->workflow->resolveStartContext($config, $onVerbose);
+        $context = $this->scrapeWorkflow->resolveStartContext($config, $onVerbose);
         if (null === $context) {
             $this->io->error('State-файл не найден. Запустите без --resume.');
 
@@ -169,7 +217,7 @@ class ScrapePartnersCommand extends Command
             $this->io->section('Парсинг партнёров...');
         }
 
-        $progressBar = $this->createProgressBar($lastPage * $partnersPerPage, count($processedNumbers));
+        $progressBar = $this->createScrapeProgressBar($lastPage * $partnersPerPage, count($processedNumbers));
 
         $onProgress = function (string $event, int $value) use ($progressBar): void {
             match ($event) {
@@ -180,7 +228,7 @@ class ScrapePartnersCommand extends Command
             };
         };
 
-        $result = $this->workflow->run(
+        $result = $this->scrapeWorkflow->run(
             $config,
             $startPage,
             $lastPage,
@@ -191,7 +239,7 @@ class ScrapePartnersCommand extends Command
         return $this->finishScrape($config->outputFile, $progressBar, $result);
     }
 
-    private function createProgressBar(int $total, int $alreadyProcessed): ?ProgressBar
+    private function createScrapeProgressBar(int $total, int $alreadyProcessed): ?ProgressBar
     {
         if ($this->output->getVerbosity() < OutputInterface::VERBOSITY_NORMAL) {
             return null;
@@ -202,6 +250,20 @@ class ScrapePartnersCommand extends Command
         $progressBar->setMessage('', 'page');
         $progressBar->setMessage('', 'partner');
         $progressBar->advance($alreadyProcessed);
+
+        return $progressBar;
+    }
+
+    private function createUpdateProgressBar(int $total): ?ProgressBar
+    {
+        if ($this->output->getVerbosity() < OutputInterface::VERBOSITY_NORMAL) {
+            return null;
+        }
+
+        $progressBar = new ProgressBar($this->output, $total);
+        $progressBar->setFormat(' %current%/%max% [%bar%] %percent:3s%% | Партнёр: %partner%');
+        $progressBar->setMessage('', 'partner');
+        $progressBar->start();
 
         return $progressBar;
     }
@@ -232,7 +294,7 @@ class ScrapePartnersCommand extends Command
             return Command::FAILURE;
         }
 
-        $this->workflow->complete($outputFile);
+        $this->scrapeWorkflow->complete($outputFile);
 
         if ($this->output->getVerbosity() >= OutputInterface::VERBOSITY_NORMAL) {
             $this->io->success(sprintf('Парсинг завершён. Обработано партнёров: %d', $result->totalProcessed));
@@ -247,5 +309,29 @@ class ScrapePartnersCommand extends Command
         }
 
         return Command::SUCCESS;
+    }
+
+    private function finishUpdate(?ProgressBar $progressBar, ScrapeResult $result): int
+    {
+        $progressBar?->finish();
+        if ($this->output->getVerbosity() >= OutputInterface::VERBOSITY_NORMAL) {
+            $this->io->newLine(2);
+
+            if ($result->banDetected) {
+                $this->io->warning(sprintf(
+                    'Обнаружена блокировка. Обновлено: %d, ошибок: %d',
+                    $result->totalProcessed,
+                    $result->totalEmptyPages,
+                ));
+            } else {
+                $this->io->success(sprintf(
+                    'Обновлено: %d, ошибок: %d',
+                    $result->totalProcessed,
+                    $result->totalEmptyPages,
+                ));
+            }
+        }
+
+        return $result->banDetected ? Command::FAILURE : Command::SUCCESS;
     }
 }
